+++ component/src/main/java/io/siddhi/extension/execution/unique/UniqueEverWindowProcessor.java
@@ -16,14 +16,13 @@
  * under the License.
  */
 
-package org.wso2.extension.siddhi.execution.unique;
+package io.siddhi.extension.execution.unique;
 
 import io.siddhi.annotation.Example;
 import io.siddhi.annotation.Extension;
 import io.siddhi.annotation.Parameter;
 import io.siddhi.annotation.util.DataType;
 import io.siddhi.core.config.SiddhiQueryContext;
-import io.siddhi.core.event.ComplexEvent;
 import io.siddhi.core.event.ComplexEventChunk;
 import io.siddhi.core.event.state.StateEvent;
 import io.siddhi.core.event.stream.MetaStreamEvent;
@@ -56,44 +55,62 @@ import java.util.concurrent.ConcurrentMap;
 import static java.util.Collections.singletonMap;
 
 /**
- * class representing unique first window processor implementation.
+ * This is Unique Ever Window Processor implementation.
  */
 
 @Extension(
-        name = "first",
+        name = "ever",
         namespace = "unique",
-        description = "This is a window that holds only the first set of unique events"
-                + " according to the unique key parameter."
-                + " When a new event arrives with a key that is already in the window,"
-                + " that event is not processed by the window.",
+        description = "This is a window that is updated with the latest events based on a unique key parameter."
+                + " When a new event arrives with the same value for the unique key parameter"
+                + " as the existing event, the existing event expires, "
+                + "and is replaced with the latest one.",
 
         parameters = {
                 @Parameter(name = "unique.key",
                         description = "The attribute that should be checked for uniqueness."
-                                + " If there is more than one parameter to check for uniqueness,"
-                                + " it can be specified as an array separated by commas.",
+                                + "If multiple attributes need to be checked, we can specify them "
+                                + "as a comma-separated list.",
                         type = {DataType.INT, DataType.LONG, DataType.FLOAT,
                                 DataType.BOOL, DataType.DOUBLE}),
         },
         examples = {
                 @Example(
-                        syntax = "define stream LoginEvents (timeStamp long, ip string);\n" +
-                                "from LoginEvents#window.unique:first(ip)\n" +
-                                "insert into UniqueIps ;",
+                        syntax = "define stream LoginEvents (timeStamp long, ip string) ;\n" +
+                                "from LoginEvents#window.unique:ever(ip)\n" +
+                                "select count(ip) as ipCount, ip \n" +
+                                "insert all events into UniqueIps  ;",
 
-                        description = "This returns the first set of unique items that arrive from the " +
-                                "'LoginEvents' stream,"
-                                + " and returns them to the 'UniqueIps' stream."
-                                + " The unique events are only those with a unique value for the 'ip' attribute."
+                        description = "The above query determines the latest events that have arrived "
+                                + "from the 'LoginEvents' stream, based on the 'ip' attribute. "
+                                + "At a given time, all the events held in the window should have a unique value "
+                                + "for the ip attribute. All the processed events are directed "
+                                + "to the 'UniqueIps' output stream with 'ip' and 'ipCount' attributes."
 
+                ),
+                @Example(
+                        syntax = "define stream LoginEvents (timeStamp long, ip string , id string) ;\n" +
+                                "from LoginEvents#window.unique:ever(ip, id)\n" +
+                                "select count(ip) as ipCount, ip , id \n" +
+                                "insert expired events into UniqueIps  ;",
+
+                        description = "This query determines the latest events to be included in the window "
+                                + "based on the ip and id attributes. When the 'LoginEvents' event stream receives"
+                                + " a new event of which the combination of values for the ip and id attributes "
+                                + "matches that of an existing event in the window, the existing event expires"
+                                + " and it is replaced with the new event. The expired events "
+                                + "which have been expired"
+                                + " as a result of being replaced by a newer event"
+                                + " are directed to the 'uniqueIps' output stream."
                 )
         }
 )
 
-public class UniqueFirstWindowProcessor extends WindowProcessor<UniqueFirstWindowProcessor.ExtensionState>
+public class UniqueEverWindowProcessor extends WindowProcessor<UniqueEverWindowProcessor.ExtensionState>
         implements FindableProcessor {
-    private ConcurrentMap<String, StreamEvent> map = new ConcurrentHashMap<String, StreamEvent>();
-    private ExpressionExecutor[] uniqueExpressionExecutors;
+    private ConcurrentMap<String, StreamEvent> map = new ConcurrentHashMap<>();
+    private ExpressionExecutor[] expressionExecutors;
+
 
     @Override
     protected StateFactory<ExtensionState> init(MetaStreamEvent metaStreamEvent, AbstractDefinition inputDefinition,
@@ -102,7 +119,7 @@ public class UniqueFirstWindowProcessor extends WindowProcessor<UniqueFirstWindo
                                                 StreamEventClonerHolder streamEventClonerHolder,
                                                 boolean outputExpectsExpiredEvents, boolean findToBeExecuted,
                                                 SiddhiQueryContext siddhiQueryContext) {
-        uniqueExpressionExecutors = attributeExpressionExecutors;
+        expressionExecutors = attributeExpressionExecutors;
         return () -> new ExtensionState();
     }
 
@@ -111,16 +128,23 @@ public class UniqueFirstWindowProcessor extends WindowProcessor<UniqueFirstWindo
                                      StreamEventCloner streamEventCloner, ComplexEventPopulater complexEventPopulater,
                                      ExtensionState state) {
         synchronized (this) {
-            while (streamEventChunk.hasNext()) {
-                StreamEvent streamEvent = streamEventChunk.next();
+            long currentTime = siddhiQueryContext.getSiddhiAppContext().getTimestampGenerator().currentTime();
 
+            StreamEvent streamEvent = streamEventChunk.getFirst();
+            streamEventChunk.clear();
+            while (streamEvent != null) {
                 StreamEvent clonedEvent = streamEventCloner.copyStreamEvent(streamEvent);
                 clonedEvent.setType(StreamEvent.Type.EXPIRED);
 
-                ComplexEvent oldEvent = map.putIfAbsent(generateKey(clonedEvent), clonedEvent);
+                StreamEvent oldEvent = map.put(generateKey(clonedEvent), clonedEvent);
                 if (oldEvent != null) {
-                    streamEventChunk.remove();
+                    oldEvent.setTimestamp(currentTime);
+                    streamEventChunk.add(oldEvent);
                 }
+                StreamEvent next = streamEvent.getNext();
+                streamEvent.setNext(null);
+                streamEventChunk.add(streamEvent);
+                streamEvent = next;
             }
         }
         nextProcessor.process(streamEventChunk);
@@ -130,6 +154,7 @@ public class UniqueFirstWindowProcessor extends WindowProcessor<UniqueFirstWindo
     public void start() {
         //Do nothing
     }
+
 
     @Override
     public void stop() {
@@ -150,21 +175,15 @@ public class UniqueFirstWindowProcessor extends WindowProcessor<UniqueFirstWindo
 
         @Override
         public Map<String, Object> snapshot() {
-            return singletonMap("map", UniqueFirstWindowProcessor.this.map);
+            return singletonMap("map", UniqueEverWindowProcessor.this.map);
         }
 
         @Override
-        public void restore(Map<String, Object> map) {
-            UniqueFirstWindowProcessor.this.map = (ConcurrentMap<String, StreamEvent>) map.get("map");
+        public void restore(Map<String, Object> state) {
+            synchronized (UniqueEverWindowProcessor.this) {
+                UniqueEverWindowProcessor.this.map = (ConcurrentMap<String, StreamEvent>) state.get("map");
+            }
         }
-    }
-
-    private String generateKey(StreamEvent event) {
-        StringBuilder stringBuilder = new StringBuilder();
-        for (ExpressionExecutor executor : uniqueExpressionExecutors) {
-            stringBuilder.append(executor.execute(event));
-        }
-        return stringBuilder.toString();
     }
 
     @Override
@@ -180,8 +199,16 @@ public class UniqueFirstWindowProcessor extends WindowProcessor<UniqueFirstWindo
     @Override
     public CompiledCondition compileCondition(Expression expression, MatchingMetaInfoHolder matchingMetaInfoHolder,
                                               List<VariableExpressionExecutor> variableExpressionExecutors,
-                                              Map<String, Table> eventTableMap, SiddhiQueryContext siddhiQueryContext) {
-        return OperatorParser.constructOperator(this.map.values(), expression, matchingMetaInfoHolder,
-                variableExpressionExecutors, eventTableMap, siddhiQueryContext);
+                                              Map<String, Table> tableMap, SiddhiQueryContext siddhiQueryContext) {
+        return OperatorParser.constructOperator(map.values(), expression, matchingMetaInfoHolder,
+                variableExpressionExecutors, tableMap, siddhiQueryContext);
+    }
+
+    private String generateKey(StreamEvent event) {
+        StringBuilder stringBuilder = new StringBuilder();
+        for (ExpressionExecutor executor : expressionExecutors) {
+            stringBuilder.append(executor.execute(event));
+        }
+        return stringBuilder.toString();
     }
 }
