+++ component/src/main/java/io/siddhi/extension/execution/unique/UniqueFirstTimeBatchWindowProcessor.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.extension.siddhi.execution.unique;
+package io.siddhi.extension.execution.unique;
 
 import io.siddhi.annotation.Example;
 import io.siddhi.annotation.Extension;
