+++ component/src/main/java/io/siddhi/extension/execution/unique/UniqueFirstWindowProcessor.java
@@ -16,13 +16,14 @@
  * under the License.
  */
 
-package org.wso2.extension.siddhi.execution.unique;
+package io.siddhi.extension.execution.unique;
 
 import io.siddhi.annotation.Example;
 import io.siddhi.annotation.Extension;
 import io.siddhi.annotation.Parameter;
 import io.siddhi.annotation.util.DataType;
 import io.siddhi.core.config.SiddhiQueryContext;
+import io.siddhi.core.event.ComplexEvent;
 import io.siddhi.core.event.ComplexEventChunk;
 import io.siddhi.core.event.state.StateEvent;
 import io.siddhi.core.event.stream.MetaStreamEvent;
@@ -55,62 +56,44 @@ import java.util.concurrent.ConcurrentMap;
 import static java.util.Collections.singletonMap;
 
 /**
- * This is Unique Ever Window Processor implementation.
+ * class representing unique first window processor implementation.
  */
 
 @Extension(
-        name = "ever",
+        name = "first",
         namespace = "unique",
-        description = "This is a window that is updated with the latest events based on a unique key parameter."
-                + " When a new event arrives with the same value for the unique key parameter"
-                + " as the existing event, the existing event expires, "
-                + "and is replaced with the latest one.",
+        description = "This is a window that holds only the first set of unique events"
+                + " according to the unique key parameter."
+                + " When a new event arrives with a key that is already in the window,"
+                + " that event is not processed by the window.",
 
         parameters = {
                 @Parameter(name = "unique.key",
                         description = "The attribute that should be checked for uniqueness."
-                                + "If multiple attributes need to be checked, we can specify them "
-                                + "as a comma-separated list.",
+                                + " If there is more than one parameter to check for uniqueness,"
+                                + " it can be specified as an array separated by commas.",
                         type = {DataType.INT, DataType.LONG, DataType.FLOAT,
                                 DataType.BOOL, DataType.DOUBLE}),
         },
         examples = {
                 @Example(
-                        syntax = "define stream LoginEvents (timeStamp long, ip string) ;\n" +
-                                "from LoginEvents#window.unique:ever(ip)\n" +
-                                "select count(ip) as ipCount, ip \n" +
-                                "insert all events into UniqueIps  ;",
+                        syntax = "define stream LoginEvents (timeStamp long, ip string);\n" +
+                                "from LoginEvents#window.unique:first(ip)\n" +
+                                "insert into UniqueIps ;",
 
-                        description = "The above query determines the latest events that have arrived "
-                                + "from the 'LoginEvents' stream, based on the 'ip' attribute. "
-                                + "At a given time, all the events held in the window should have a unique value "
-                                + "for the ip attribute. All the processed events are directed "
-                                + "to the 'UniqueIps' output stream with 'ip' and 'ipCount' attributes."
+                        description = "This returns the first set of unique items that arrive from the " +
+                                "'LoginEvents' stream,"
+                                + " and returns them to the 'UniqueIps' stream."
+                                + " The unique events are only those with a unique value for the 'ip' attribute."
 
-                ),
-                @Example(
-                        syntax = "define stream LoginEvents (timeStamp long, ip string , id string) ;\n" +
-                                "from LoginEvents#window.unique:ever(ip, id)\n" +
-                                "select count(ip) as ipCount, ip , id \n" +
-                                "insert expired events into UniqueIps  ;",
-
-                        description = "This query determines the latest events to be included in the window "
-                                + "based on the ip and id attributes. When the 'LoginEvents' event stream receives"
-                                + " a new event of which the combination of values for the ip and id attributes "
-                                + "matches that of an existing event in the window, the existing event expires"
-                                + " and it is replaced with the new event. The expired events "
-                                + "which have been expired"
-                                + " as a result of being replaced by a newer event"
-                                + " are directed to the 'uniqueIps' output stream."
                 )
         }
 )
 
-public class UniqueEverWindowProcessor extends WindowProcessor<UniqueEverWindowProcessor.ExtensionState>
+public class UniqueFirstWindowProcessor extends WindowProcessor<UniqueFirstWindowProcessor.ExtensionState>
         implements FindableProcessor {
-    private ConcurrentMap<String, StreamEvent> map = new ConcurrentHashMap<>();
-    private ExpressionExecutor[] expressionExecutors;
-
+    private ConcurrentMap<String, StreamEvent> map = new ConcurrentHashMap<String, StreamEvent>();
+    private ExpressionExecutor[] uniqueExpressionExecutors;
 
     @Override
     protected StateFactory<ExtensionState> init(MetaStreamEvent metaStreamEvent, AbstractDefinition inputDefinition,
@@ -119,7 +102,7 @@ public class UniqueEverWindowProcessor extends WindowProcessor<UniqueEverWindowP
                                                 StreamEventClonerHolder streamEventClonerHolder,
                                                 boolean outputExpectsExpiredEvents, boolean findToBeExecuted,
                                                 SiddhiQueryContext siddhiQueryContext) {
-        expressionExecutors = attributeExpressionExecutors;
+        uniqueExpressionExecutors = attributeExpressionExecutors;
         return () -> new ExtensionState();
     }
 
@@ -128,23 +111,16 @@ public class UniqueEverWindowProcessor extends WindowProcessor<UniqueEverWindowP
                                      StreamEventCloner streamEventCloner, ComplexEventPopulater complexEventPopulater,
                                      ExtensionState state) {
         synchronized (this) {
-            long currentTime = siddhiQueryContext.getSiddhiAppContext().getTimestampGenerator().currentTime();
+            while (streamEventChunk.hasNext()) {
+                StreamEvent streamEvent = streamEventChunk.next();
 
-            StreamEvent streamEvent = streamEventChunk.getFirst();
-            streamEventChunk.clear();
-            while (streamEvent != null) {
                 StreamEvent clonedEvent = streamEventCloner.copyStreamEvent(streamEvent);
                 clonedEvent.setType(StreamEvent.Type.EXPIRED);
 
-                StreamEvent oldEvent = map.put(generateKey(clonedEvent), clonedEvent);
+                ComplexEvent oldEvent = map.putIfAbsent(generateKey(clonedEvent), clonedEvent);
                 if (oldEvent != null) {
-                    oldEvent.setTimestamp(currentTime);
-                    streamEventChunk.add(oldEvent);
+                    streamEventChunk.remove();
                 }
-                StreamEvent next = streamEvent.getNext();
-                streamEvent.setNext(null);
-                streamEventChunk.add(streamEvent);
-                streamEvent = next;
             }
         }
         nextProcessor.process(streamEventChunk);
@@ -154,7 +130,6 @@ public class UniqueEverWindowProcessor extends WindowProcessor<UniqueEverWindowP
     public void start() {
         //Do nothing
     }
-
 
     @Override
     public void stop() {
@@ -175,15 +150,21 @@ public class UniqueEverWindowProcessor extends WindowProcessor<UniqueEverWindowP
 
         @Override
         public Map<String, Object> snapshot() {
-            return singletonMap("map", UniqueEverWindowProcessor.this.map);
+            return singletonMap("map", UniqueFirstWindowProcessor.this.map);
         }
 
         @Override
-        public void restore(Map<String, Object> state) {
-            synchronized (UniqueEverWindowProcessor.this) {
-                UniqueEverWindowProcessor.this.map = (ConcurrentMap<String, StreamEvent>) state.get("map");
-            }
+        public void restore(Map<String, Object> map) {
+            UniqueFirstWindowProcessor.this.map = (ConcurrentMap<String, StreamEvent>) map.get("map");
         }
+    }
+
+    private String generateKey(StreamEvent event) {
+        StringBuilder stringBuilder = new StringBuilder();
+        for (ExpressionExecutor executor : uniqueExpressionExecutors) {
+            stringBuilder.append(executor.execute(event));
+        }
+        return stringBuilder.toString();
     }
 
     @Override
@@ -199,16 +180,8 @@ public class UniqueEverWindowProcessor extends WindowProcessor<UniqueEverWindowP
     @Override
     public CompiledCondition compileCondition(Expression expression, MatchingMetaInfoHolder matchingMetaInfoHolder,
                                               List<VariableExpressionExecutor> variableExpressionExecutors,
-                                              Map<String, Table> tableMap, SiddhiQueryContext siddhiQueryContext) {
-        return OperatorParser.constructOperator(map.values(), expression, matchingMetaInfoHolder,
-                variableExpressionExecutors, tableMap, siddhiQueryContext);
-    }
-
-    private String generateKey(StreamEvent event) {
-        StringBuilder stringBuilder = new StringBuilder();
-        for (ExpressionExecutor executor : expressionExecutors) {
-            stringBuilder.append(executor.execute(event));
-        }
-        return stringBuilder.toString();
+                                              Map<String, Table> eventTableMap, SiddhiQueryContext siddhiQueryContext) {
+        return OperatorParser.constructOperator(this.map.values(), expression, matchingMetaInfoHolder,
+                variableExpressionExecutors, eventTableMap, siddhiQueryContext);
     }
 }
