+++ component/src/main/java/io/siddhi/extension/execution/unique/UniqueTimeBatchWindowProcessor.java
@@ -1,12 +1,12 @@
 /*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.wso2.extension.siddhi.execution.unique;
+package io.siddhi.extension.execution.unique;
 
 import io.siddhi.annotation.Example;
 import io.siddhi.annotation.Extension;
@@ -57,18 +57,17 @@ import java.util.List;
 import java.util.Map;
 
 /**
- * class representing Unique Time Length Batch window processor implementation.
+ * class representing unique time batch window processor implementation.
  */
 
 @Extension(
-        name = "timeLengthBatch",
+        name = "timeBatch",
         namespace = "unique",
-        description = "This is a batch or tumbling time length window that is updated "
-                + "with the latest events based on a unique key parameter. The window tumbles "
-                + "upon the elapse of the time window, or when a number of unique events have arrived."
-                + " If a new event that arrives within the period of the window "
-                + "has a value for the key parameter which matches the value of an existing event, "
-                + "the existing event expires and it is replaced by the new event. ",
+        description = "This is a batch (tumbling) time window that is updated "
+                + "with the latest events based on a unique key parameter."
+                + " If a new event that arrives within the time period of a window"
+                + "has a value for the key parameter which matches that of an existing event, "
+                + "the existing event expires and it is replaced by the latest event. ",
         parameters = {
                 @Parameter(name = "unique.key",
                         description = "The attribute that should be checked for uniqueness.",
@@ -76,7 +75,7 @@ import java.util.Map;
                                 DataType.BOOL, DataType.DOUBLE}),
 
                 @Parameter(name = "window.time",
-                        description = "The sliding time period for which the window should hold the events.",
+                        description = "The tumbling time period for which the window should hold events.",
                         type = {DataType.INT, DataType.LONG}),
 
                 @Parameter(name = "start.time",
@@ -84,16 +83,12 @@ import java.util.Map;
                                 " window at a time different to the standard time.",
                         type = {DataType.INT, DataType.LONG},
                         optional = true,
-                        defaultValue = "Timestamp of first event"),
-
-                @Parameter(name = "window.length",
-                        description = "The number of events the window should tumble.",
-                        type = {DataType.INT})
+                        defaultValue = "Timestamp of first event")
         },
         examples = {
                 @Example(
                         syntax = "define stream CseEventStream (symbol string, price float, volume int)\n\n" +
-                                "from CseEventStream#window.unique:timeLengthBatch(symbol, 1 sec, 20)\n" +
+                                "from CseEventStream#window.unique:timeBatch(symbol, 1 sec)\n" +
                                 "select symbol, price, volume\n" +
                                 "insert all events into OutputStream ;",
 
@@ -104,12 +99,10 @@ import java.util.Map;
         }
 )
 
-public class UniqueTimeLengthBatchWindowProcessor
-        extends WindowProcessor<UniqueTimeLengthBatchWindowProcessor.ExtensionState> implements SchedulingProcessor,
-        FindableProcessor {
+public class UniqueTimeBatchWindowProcessor extends WindowProcessor<UniqueTimeBatchWindowProcessor.ExtensionState>
+        implements SchedulingProcessor, FindableProcessor {
 
     private long timeInMilliSeconds;
-    private long length;
     private long nextEmitTime = -1;
     private ComplexEventChunk<StreamEvent> eventsToBeExpired = null;
     private Map<Object, StreamEvent> uniqueEventMap = new HashMap<>();
@@ -117,7 +110,7 @@ public class UniqueTimeLengthBatchWindowProcessor
     private boolean isStartTimeEnabled = false;
     private long startTime = 0;
     private ExpressionExecutor uniqueKeyExpressionExecutor;
-    private boolean eventSent = false;
+
 
     @Override
     protected StateFactory<ExtensionState> init(MetaStreamEvent metaStreamEvent, AbstractDefinition inputDefinition,
@@ -127,7 +120,7 @@ public class UniqueTimeLengthBatchWindowProcessor
                                                 boolean outputExpectsExpiredEvents, boolean findToBeExecuted,
                                                 SiddhiQueryContext siddhiQueryContext) {
         this.eventsToBeExpired = new ComplexEventChunk<>(false);
-        if (attributeExpressionExecutors.length == 3) {
+        if (attributeExpressionExecutors.length == 2) {
             this.uniqueKeyExpressionExecutor = attributeExpressionExecutors[0];
             if (attributeExpressionExecutors[1] instanceof ConstantExpressionExecutor) {
                 if (attributeExpressionExecutors[1].getReturnType() == Attribute.Type.INT) {
@@ -138,32 +131,15 @@ public class UniqueTimeLengthBatchWindowProcessor
                             .getValue();
                 } else {
                     throw new SiddhiAppValidationException(
-                            "Unique Time Length Batch window's parameter time should be either int, long or time, " +
-                                    "but found " + attributeExpressionExecutors[1].getReturnType());
+                            "Unique Time Batch window's parameter " + "time should be either"
+                                    + "int or long, but found " + attributeExpressionExecutors[1].getReturnType());
                 }
             } else {
-                throw new SiddhiAppValidationException("Unique Time Length Batch window should have constant "
+                throw new SiddhiAppValidationException("Unique Time Batch window should have constant "
                         + "for time parameter but found a dynamic attribute " + attributeExpressionExecutors[1]
                         .getClass().getCanonicalName());
             }
-            if (attributeExpressionExecutors[2] instanceof ConstantExpressionExecutor) {
-                if (attributeExpressionExecutors[2].getReturnType() == Attribute.Type.INT) {
-                    length = (Integer) ((ConstantExpressionExecutor) attributeExpressionExecutors[2])
-                            .getValue();
-                } else if (attributeExpressionExecutors[2].getReturnType() == Attribute.Type.LONG) {
-                    length = (Long) ((ConstantExpressionExecutor) attributeExpressionExecutors[2])
-                            .getValue();
-                } else {
-                    throw new SiddhiAppValidationException(
-                            "Unique Time Length Batch window's parameter " + "length should be either"
-                                    + "int or long, but found " + attributeExpressionExecutors[2].getReturnType());
-                }
-            } else {
-                throw new SiddhiAppValidationException("Unique Time Length Batch window should have constant "
-                        + "for length parameter but found a dynamic attribute " + attributeExpressionExecutors[2]
-                        .getClass().getCanonicalName());
-            }
-        } else if (attributeExpressionExecutors.length == 4) {
+        } else if (attributeExpressionExecutors.length == 3) {
             this.uniqueKeyExpressionExecutor = attributeExpressionExecutors[0];
             if (attributeExpressionExecutors[1] instanceof ConstantExpressionExecutor) {
                 if (attributeExpressionExecutors[1].getReturnType() == Attribute.Type.INT) {
@@ -174,12 +150,11 @@ public class UniqueTimeLengthBatchWindowProcessor
                             .getValue();
                 } else {
                     throw new SiddhiAppValidationException(
-                            "UniqueTimeLengthBatch window's parameter time should be either" +
-                                    " int or long, but found "
+                            "UniqueTimeBatch window's parameter time should be either" + " int or long, but found "
                                     + attributeExpressionExecutors[1].getReturnType());
                 }
             } else {
-                throw new SiddhiAppValidationException("Unique Time Length Batch window should have constant "
+                throw new SiddhiAppValidationException("Unique Time Batch window should have constant "
                         + "for time parameter but found a dynamic attribute " + attributeExpressionExecutors[1]
                         .getClass().getCanonicalName());
             }
@@ -194,35 +169,18 @@ public class UniqueTimeLengthBatchWindowProcessor
                     startTime = Long.parseLong(
                             String.valueOf(((ConstantExpressionExecutor) attributeExpressionExecutors[2]).getValue()));
                 } else {
-                    throw new SiddhiAppValidationException("Expected either int or long type for " +
-                            "UniqueTimeLengthBatch window's start time parameter, but found "
+                    throw new SiddhiAppValidationException("Expected either "
+                            + "int or long type for UniqueTimeBatch window's start time parameter, but found "
                             + attributeExpressionExecutors[2].getReturnType());
                 }
             } else {
-                throw new SiddhiAppValidationException("Unique Time Length Batch window should have constant "
+                throw new SiddhiAppValidationException("Unique Time Batch window should have constant "
                         + "for time parameter but found a dynamic attribute " + attributeExpressionExecutors[2]
                         .getReturnType());
             }
-            if (attributeExpressionExecutors[3] instanceof ConstantExpressionExecutor) {
-                if (attributeExpressionExecutors[3].getReturnType() == Attribute.Type.INT) {
-                    length = (Integer) ((ConstantExpressionExecutor) attributeExpressionExecutors[3])
-                            .getValue();
-                } else if (attributeExpressionExecutors[3].getReturnType() == Attribute.Type.LONG) {
-                    length = (Long) ((ConstantExpressionExecutor) attributeExpressionExecutors[3])
-                            .getValue();
-                } else {
-                    throw new SiddhiAppValidationException(
-                            "Unique Time Length Batch window's parameter " + "length should be either"
-                                    + "int or long, but found " + attributeExpressionExecutors[3].getReturnType());
-                }
-            } else {
-                throw new SiddhiAppValidationException("Unique Time Length Batch window should have constant "
-                        + "for length parameter but found a dynamic attribute " + attributeExpressionExecutors[3]
-                        .getClass().getCanonicalName());
-            }
         } else {
             throw new SiddhiAppValidationException(
-                    "Unique Time Length Batch window should " + "only have three or four parameters. " + "but found "
+                    "Unique Time Batch window should " + "only have two or three parameters. " + "but found "
                             + attributeExpressionExecutors.length + " input attributes");
         }
         return () -> new ExtensionState();
@@ -244,23 +202,17 @@ public class UniqueTimeLengthBatchWindowProcessor
                     scheduler.notifyAt(nextEmitTime);
                 }
             }
-            boolean sendEventsByTime = false;
-            boolean sendEventsByLength = false;
+            boolean sendEvents;
             if (currentTime >= nextEmitTime) {
                 nextEmitTime += timeInMilliSeconds;
+
                 if (scheduler != null) {
                     scheduler.notifyAt(nextEmitTime);
                 }
-                if (eventSent) { // reset on next batch
-                    eventSent = false;
-                    streamEventChunk.clear();
-                    return;
-                }
-                sendEventsByTime = true;
-            }
-            if (eventSent) { //skip events till next batch
-                streamEventChunk.clear();
-                return;
+
+                sendEvents = true;
+            } else {
+                sendEvents = false;
             }
             while (streamEventChunk.hasNext()) {
                 StreamEvent streamEvent = streamEventChunk.next();
@@ -269,17 +221,39 @@ public class UniqueTimeLengthBatchWindowProcessor
                 }
                 StreamEvent clonedStreamEvent = streamEventCloner.copyStreamEvent(streamEvent);
                 addUniqueEvent(uniqueEventMap, uniqueKeyExpressionExecutor, clonedStreamEvent);
-                if (uniqueEventMap.size() == length) {
-                    sendEventsByLength = true; // emitting batch based on length
-                    break;
-                }
             }
             streamEventChunk.clear();
-            if (sendEventsByTime || sendEventsByLength) {
-                sendEvents(streamEventChunk, streamEventCloner, currentTime, state);
-            }
-            if (sendEventsByLength) {
-                eventSent = true; // making events to skip till next time batch
+            if (sendEvents) {
+                for (StreamEvent event : uniqueEventMap.values()) {
+                    event.setTimestamp(currentTime);
+                    state.currentEventChunk.add(event);
+                }
+                uniqueEventMap.clear();
+                if (eventsToBeExpired.getFirst() != null) {
+                    while (eventsToBeExpired.hasNext()) {
+                        StreamEvent expiredEvent = eventsToBeExpired.next();
+                        expiredEvent.setTimestamp(currentTime);
+                    }
+                    streamEventChunk.add(eventsToBeExpired.getFirst());
+                }
+                eventsToBeExpired.clear();
+                if (state.currentEventChunk.getFirst() != null) {
+                    // add reset event in front of current events
+                    streamEventChunk.add(state.resetEvent);
+                    state.currentEventChunk.reset();
+                    while (state.currentEventChunk.hasNext()) {
+                        StreamEvent streamEvent = state.currentEventChunk.next();
+                        StreamEvent eventClonedForMap = streamEventCloner.copyStreamEvent(streamEvent);
+                        eventClonedForMap.setType(StreamEvent.Type.EXPIRED);
+                        this.eventsToBeExpired.add(eventClonedForMap);
+                    }
+                    if (state.currentEventChunk.getFirst() != null) {
+                        state.resetEvent = streamEventCloner.copyStreamEvent(state.currentEventChunk.getFirst());
+                        state.resetEvent.setType(ComplexEvent.Type.RESET);
+                        streamEventChunk.add(state.currentEventChunk.getFirst());
+                    }
+                }
+                state.currentEventChunk.clear();
             }
         }
         if (streamEventChunk.getFirst() != null) {
@@ -287,40 +261,6 @@ public class UniqueTimeLengthBatchWindowProcessor
             nextProcessor.process(streamEventChunk);
             streamEventChunk.setBatch(false);
         }
-    }
-
-    private void sendEvents(ComplexEventChunk<StreamEvent> streamEventChunk, StreamEventCloner streamEventCloner,
-                            long currentTime, ExtensionState state) {
-        for (StreamEvent event : uniqueEventMap.values()) {
-            event.setTimestamp(currentTime);
-            state.currentEventChunk.add(event);
-        }
-        uniqueEventMap.clear();
-        if (eventsToBeExpired.getFirst() != null) {
-            while (eventsToBeExpired.hasNext()) {
-                StreamEvent expiredEvent = eventsToBeExpired.next();
-                expiredEvent.setTimestamp(currentTime);
-            }
-            streamEventChunk.add(eventsToBeExpired.getFirst());
-        }
-        eventsToBeExpired.clear();
-        if (state.currentEventChunk.getFirst() != null) {
-            // add reset event in front of current events
-            streamEventChunk.add(state.resetEvent);
-            state.currentEventChunk.reset();
-            while (state.currentEventChunk.hasNext()) {
-                StreamEvent streamEvent = state.currentEventChunk.next();
-                StreamEvent eventClonedForMap = streamEventCloner.copyStreamEvent(streamEvent);
-                eventClonedForMap.setType(StreamEvent.Type.EXPIRED);
-                this.eventsToBeExpired.add(eventClonedForMap);
-            }
-            if (state.currentEventChunk.getFirst() != null) {
-                state.resetEvent = streamEventCloner.copyStreamEvent(state.currentEventChunk.getFirst());
-                state.resetEvent.setType(ComplexEvent.Type.RESET);
-                streamEventChunk.add(state.currentEventChunk.getFirst());
-            }
-        }
-        state.currentEventChunk.clear();
     }
 
 
@@ -368,8 +308,8 @@ public class UniqueTimeLengthBatchWindowProcessor
 
     class ExtensionState extends State {
 
-        private ComplexEventChunk<StreamEvent> currentEventChunk = new ComplexEventChunk<>(false);
         private StreamEvent resetEvent = null;
+        private ComplexEventChunk<StreamEvent> currentEventChunk = new ComplexEventChunk<>(false);
 
         @Override
         public boolean canDestroy() {
@@ -412,7 +352,8 @@ public class UniqueTimeLengthBatchWindowProcessor
     public StreamEvent find(StateEvent matchingEvent, CompiledCondition compiledCondition) {
         if (compiledCondition instanceof Operator) {
             return ((Operator) compiledCondition).find(matchingEvent, eventsToBeExpired,
-                    streamEventClonerHolder.getStreamEventCloner());
+                    streamEventClonerHolder.getStreamEventCloner()
+            );
         } else {
             return null;
         }
