+++ component/src/test/java/io/siddhi/extension/execution/unique/UniqueFirstWindowTestCase.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.extension.siddhi.execution.unique;
+package io.siddhi.extension.execution.unique;
 
 import io.siddhi.core.SiddhiAppRuntime;
 import io.siddhi.core.SiddhiManager;
@@ -37,43 +37,45 @@ import org.testng.annotations.Test;
 import java.util.concurrent.atomic.AtomicInteger;
 
 /**
- * class representing unique ever window processor test case.
+ * class representing unique first window test case.
  */
-public class UniqueEverWindowTestCase {
-    private static final Logger log = Logger.getLogger(UniqueEverWindowTestCase.class);
-    private long value;
+
+public class UniqueFirstWindowTestCase {
+    private static final Logger log = Logger.getLogger(UniqueFirstWindowTestCase.class);
+    private int count;
     private boolean eventArrived;
     private int waitTime = 50;
     private int timeout = 30000;
+    private int lastValueRemoved;
     private AtomicInteger eventCount;
-    private int lastRemoveValue;
 
     @BeforeMethod public void init() {
-        value = 0;
+        count = 0;
         eventArrived = false;
         eventCount = new AtomicInteger(0);
-        lastRemoveValue = 0;
+        lastValueRemoved = 0;
     }
 
-    @Test public void uniqueEverWindowTest1() throws InterruptedException {
-        log.info("uniqueEverWindow test1");
+    @Test public void uniqueFirstWindowTest1() throws InterruptedException {
+        log.info("UniqueFirstWindow test1");
 
         SiddhiManager siddhiManager = new SiddhiManager();
 
         String cseEventStream = "" + "define stream LoginEvents (timeStamp long, ip string);";
-        String query = "" + "@info(name = 'query1') " + "from LoginEvents#window.unique:ever(ip) "
-                + "select count(ip) as ipCount, ip " + "insert into uniqueIps ;";
+        String query = "" + "@info(name = 'query1') " + "from LoginEvents#window.unique:first(ip) " + "select ip "
+                + "insert into uniqueIps ;";
 
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
 
         siddhiAppRuntime.addCallback("query1", new QueryCallback() {
             @Override public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
                 EventPrinter.print(timeStamp, inEvents, removeEvents);
-                for (Event event : inEvents) {
-                    eventCount.incrementAndGet();
-                }
+                eventCount.incrementAndGet();
                 if (inEvents != null) {
-                    value = (Long) inEvents[inEvents.length - 1].getData(0);
+                    count = count + inEvents.length;
+                }
+                if (removeEvents != null) {
+                    Assert.fail("Remove events emitted");
                 }
                 eventArrived = true;
             }
@@ -89,33 +91,34 @@ public class UniqueEverWindowTestCase {
         inputHandler.send(new Object[] { System.currentTimeMillis(), "192.10.1.3" });
         inputHandler.send(new Object[] { System.currentTimeMillis(), "192.10.1.5" });
 
-        SiddhiTestHelper.waitForEvents(waitTime, 5, eventCount, timeout);
-        Assert.assertEquals(eventArrived, true, "Event arrived");
-        Assert.assertEquals(value, 3, "Event max value");
+        SiddhiTestHelper.waitForEvents(waitTime, 3, eventCount, timeout);
 
+        Assert.assertEquals(eventArrived, true, "Event arrived");
+        Assert.assertEquals(count, 3, "Number of output event value");
         siddhiAppRuntime.shutdown();
+
     }
 
-    @Test public void uniqueEverWindowTest2() throws InterruptedException {
-        log.info("uniqueEverWindow test2");
+    @Test public void firstUniqueWindowTest2() throws InterruptedException {
+        log.info("UniqueFirstWindow test2");
 
         SiddhiManager siddhiManager = new SiddhiManager();
 
-        String cseEventStream = "" + "define stream LoginEvents (id String , timeStamp long, ip string);";
-        String query = "" + "@info(name = 'query1') " + "from LoginEvents#window.unique:ever(ip) "
-                + "select id, count(ip) as ipCount, ip " + "insert into uniqueIps ;";
+        String cseEventStream = "" + "define stream LoginEvents (timeStamp long, ip string);";
+        String query = "" + "@info(name = 'query1') " + "from LoginEvents#window.unique:first(ip) " + "select ip "
+                + "insert into uniqueIps ;";
 
         SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(cseEventStream + query);
 
         siddhiAppRuntime.addCallback("query1", new QueryCallback() {
             @Override public void receive(long timeStamp, Event[] inEvents, Event[] removeEvents) {
                 EventPrinter.print(timeStamp, inEvents, removeEvents);
-                for (Event event : inEvents) {
-                    eventCount.incrementAndGet();
-                }
-
+                eventCount.incrementAndGet();
                 if (inEvents != null) {
-                    value = (Long) inEvents[inEvents.length - 1].getData(1);
+                    count = count + inEvents.length;
+                }
+                if (removeEvents != null) {
+                    Assert.fail("Remove events emitted");
                 }
                 eventArrived = true;
             }
@@ -125,22 +128,22 @@ public class UniqueEverWindowTestCase {
         InputHandler inputHandler = siddhiAppRuntime.getInputHandler("LoginEvents");
         siddhiAppRuntime.start();
 
-        inputHandler.send(new Object[] { "A1", System.currentTimeMillis(), "192.10.1.3" });
-        inputHandler.send(new Object[] { "A2", System.currentTimeMillis(), "192.10.1.3" });
-        inputHandler.send(new Object[] { "A3", System.currentTimeMillis(), "192.10.1.4" });
-        inputHandler.send(new Object[] { "A4", System.currentTimeMillis(), "192.10.1.3" });
-        inputHandler.send(new Object[] { "A5", System.currentTimeMillis(), "192.10.1.5" });
+        inputHandler.send(new Object[] { System.currentTimeMillis(), "192.10.1.3" });
+        inputHandler.send(new Object[] { System.currentTimeMillis(), "192.10.1.12" });
+        inputHandler.send(new Object[] { System.currentTimeMillis(), "192.10.1.3" });
+        inputHandler.send(new Object[] { System.currentTimeMillis(), "192.10.1.3" });
+        inputHandler.send(new Object[] { System.currentTimeMillis(), "192.10.1.3" });
 
-        SiddhiTestHelper.waitForEvents(waitTime, 5, eventCount, timeout);
+        SiddhiTestHelper.waitForEvents(waitTime, 2, eventCount, timeout);
+
         Assert.assertEquals(eventArrived, true, "Event arrived");
-        Assert.assertEquals(value, 3, "Event max value");
-
+        Assert.assertEquals(count, 2, "Number of output event value");
         siddhiAppRuntime.shutdown();
     }
 
     @Test
-    public void uniqueEverWindowTest3() throws InterruptedException {
-        log.info("uniqueEverWindowTest3 - unique window query");
+    public void firstUniqueWindowTest3() throws InterruptedException {
+        log.info("firstUniqueWindowTest3 - first unique window query");
 
         PersistenceStore persistenceStore = new InMemoryPersistenceStore();
         SiddhiManager siddhiManager = new SiddhiManager();
@@ -152,7 +155,7 @@ public class UniqueEverWindowTestCase {
                 "define stream StockStream ( symbol string, price float, volume int );" +
                 "" +
                 "@info(name = 'query1')" +
-                "from StockStream[price>10]#window.unique:ever(symbol) " +
+                "from StockStream[price>10]#window.unique:first(symbol) " +
                 "select * " +
                 "insert all events into OutStream ";
 
@@ -163,13 +166,7 @@ public class UniqueEverWindowTestCase {
                 eventArrived = true;
                 eventCount.incrementAndGet();
                 for (Event inEvent : inEvents) {
-                    AssertJUnit.assertTrue("IBM".equals(inEvent.getData(0)) || "WSO2".equals(inEvent.getData(0)));
-                }
-
-                if (removeEvents != null) {
-                    for (Event removeEvent : removeEvents) {
-                        lastRemoveValue = (Integer) removeEvent.getData(2);
-                    }
+                    lastValueRemoved = (Integer) inEvent.getData(2);
                 }
             }
         };
@@ -182,17 +179,15 @@ public class UniqueEverWindowTestCase {
 
         inputHandler.send(new Object[]{"IBM", 75.6f, 100});
         inputHandler.send(new Object[]{"WSO2", 75.6f, 100});
-        inputHandler.send(new Object[]{"IBM", 75.6f, 110});
         AssertJUnit.assertTrue(eventArrived);
         Thread.sleep(500);
-        AssertJUnit.assertEquals(100, lastRemoveValue);
+        AssertJUnit.assertEquals(100, lastValueRemoved);
 
         //persisting
         executionPlanRuntime.persist();
         Thread.sleep(500);
 
-        inputHandler.send(new Object[]{"WSO2", 75.6f, 50});
-        inputHandler.send(new Object[]{"IBM", 75.6f, 50});
+        inputHandler.send(new Object[]{"MIT", 75.6f, 110});
 
         //restarting execution plan
         Thread.sleep(500);
@@ -209,30 +204,30 @@ public class UniqueEverWindowTestCase {
             Assert.fail("Error in restoring last revision");
         }
 
-        inputHandler.send(new Object[]{"IBM", 75.6f, 100});
+        inputHandler.send(new Object[]{"MIT", 75.6f, 100});
+        inputHandler.send(new Object[]{"WSO2", 75.6f, 110});
 
-        //shutdown execution plan
-        SiddhiTestHelper.waitForEvents(100, 6, eventCount, 10000);
+        SiddhiTestHelper.waitForEvents(100, 4, eventCount, 10000);
         AssertJUnit.assertEquals(true, eventArrived);
-        AssertJUnit.assertEquals(110, lastRemoveValue);
+        AssertJUnit.assertEquals(100, lastValueRemoved);
         executionPlanRuntime.shutdown();
     }
 
     @Test
-    public void uniqueEverWindowTest4() throws InterruptedException {
-        log.info("uniqueEverWindowTest3");
+    public void firstUniqueWindowTest4() throws InterruptedException {
+        log.info("firstUniqueWindowTest2");
         SiddhiManager siddhiManager = new SiddhiManager();
         String streams = "" +
                 "define stream cseEventStream (symbol string, price float, volume int); " +
                 "define stream twitterStream (user string, tweet string, company string); ";
         String query = "" +
                 "@info(name = 'query1') " +
-                "from cseEventStream#window.unique:ever(symbol) join twitterStream#window.unique:ever(user) " +
+                "from cseEventStream#window.unique:first(symbol) join twitterStream#window.unique:ever(user) " +
                 "on cseEventStream.symbol== twitterStream.company " +
                 "select cseEventStream.symbol as symbol, twitterStream.tweet, cseEventStream.price " +
                 "insert into outputStream ;";
 
-       SiddhiAppRuntime executionPlanRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+        SiddhiAppRuntime executionPlanRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
         try {
             executionPlanRuntime.addCallback("query1", new QueryCallback() {
                 @Override
@@ -254,11 +249,12 @@ public class UniqueEverWindowTestCase {
             twitterStreamHandler.send(new Object[]{"User1", "Hello World", "WSO2"});
             twitterStreamHandler.send(new Object[]{"User2", "Hello World2", "WSO2"});
             cseEventStreamHandler.send(new Object[]{"WSO2", 75.6f, 100});
-            SiddhiTestHelper.waitForEvents(100, 4, eventCount, 10000);
-            Assert.assertEquals(4, eventCount.get());
+            SiddhiTestHelper.waitForEvents(100, 2, eventCount, 10000);
+            Assert.assertEquals(2, eventCount.get());
             Assert.assertTrue(eventArrived);
         } finally {
             executionPlanRuntime.shutdown();
         }
     }
+
 }
