+++ pom.xml
@@ -25,7 +25,7 @@
         <version>5</version>
     </parent>
 
-    <groupId>org.wso2.extension.siddhi.execution.unique</groupId>
+    <groupId>io.siddhi.extension.execution.unique</groupId>
     <artifactId>siddhi-execution-unique-parent</artifactId>
     <version>5.0.0-SNAPSHOT</version>
     <packaging>pom</packaging>
